Specific Transaction Editing
The transactions can now be edited by clicking on the edit button
specific to that transaction. The edit uses AJAX so the page doesn’t
need a refresh. The new transaction info is updated on the database. A
new PHP page, update server.php, was made to retrieve the data for the
XML Request.