CSS for Login Page
Added the CSS for the login page. Also, the login page now displays the
login form only if you are not already logged in. If you are logged in,
then it asks you if you want to log out, or give you the option to go
to your home page.